fix id

+++ src/database/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from pydantic import EmailStr
-from sqlalchemy import BigInteger, Text, DateTime, func
+from sqlalchemy import BigInteger, Text, DateTime, func, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database import Base
@@ -10,7 +10,7 @@ from database import Base
 class User(Base):
     __tablename__ = "User"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[Text] = mapped_column(Text, nullable=False)
     email: Mapped[EmailStr] = mapped_column(Text, nullable=False, unique=True)
     password: Mapped[str] = mapped_column(Text, nullable=False)
